Move Tasks to ecosystems
